sriov, Improve pfs allocation function modularity
Currently the pfs allocation function gets
the node it should allocate pfs to.

In order to increase modularity,
add a parameter which determines how many pfs it should allocate to the node.
This way we would be able to use it in a modular
way in case we need to control the number of pfs
we allocate to that node.

Currently sriov actually uses just one of the pfs,
but allocates all available whitelisted pfs.
Therefore this commit doesn't affect the end result
(beside allocating just one pf by default).

Signed-off-by: Or Shoval <oshoval@redhat.com>