sriov, Fix sriov pfs scan
In case there are no sriov pfs available,
current sriov pfs scan logic would not detect it.
Bash would keep the array non empty,
and the only element would not be a valid pf
but the path itself.

Fix logic to use find, which returns an empty
output in case no valid matches found.

Since we use local keyword, the command would not
fail the script, even that we use -e.
Therefore add a check if the array is empty,
and fail fast in this case.
This will detect the problems as soon as possible
without trailing errors.

Signed-off-by: Or Shoval <oshoval@redhat.com>